requirements.txt
flask
flask_cors
opencv-python
google-generativeai
python-dotenv
torch
pillow
transformers
skimage
ultralytics
torchvision


